requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
python-magic==0.4.27
sqlparse==0.5.1
typing_extensions==4.12.2
